Add files via upload
Good morning,
I have added my scratch files as well but the important code is in Chattermill-DC.py and I have a notebook CHML-Final.
The other notebook is the one that I was experimenting. Also I didn't know if you needed day of week or day generally so I added both averages.
My preprocessing of the reviews data probably is not very sophisticated but for a couple of hours work I think is enough. Let me know if this is satisfactory. Any feedback is appreciated.